Move styling related files
