+++ src/styles/theme.js
@@ -1,5 +1,5 @@
-import { createTheme } from '@mui/material/styles'
 import { red } from '@mui/material/colors'
+import { createTheme } from '@mui/material/styles'
 
 // Create a theme instance.
 const theme = createTheme({
